requirements.txt
asgiref==3.4.1
beautifulsoup4==4.10.0
certifi==2021.10.8
charset-normalizer==2.0.12
Django==3.2.12
django-rest-framework==0.1.0
djangorestframework==3.13.1
idna==3.3
lxml==4.8.0
pytz==2022.1
requests==2.27.1
soupsieve==2.3.1
sqlparse==0.4.2
typing_extensions==4.1.1
urllib3==1.26.9
